+++ packages/dumb-worker-router/CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added paramter for object that has fetch method to be passed to `fetcher`
+- Added parameter for object that has fetch method to be passed to `fetcher`
 
 ## [0.0.6] - 2023-04-17